Moved all interfaces into their own file
