+++ PluginAPI/PluginAPI.csproj
@@ -46,6 +46,7 @@
     <Compile Include="CommandManager.cs" />
     <Compile Include="DateEvents.cs" />
     <Compile Include="ExtensionMethods.cs" />
+    <Compile Include="Interfaces.cs" />
     <Compile Include="PluginManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utility.cs" />
